# laravelApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001"
        }
    ],
    "paths": {
...
        "/api/inscripciones": {
            "get": {
                "summary": "Mostrar todas las inscripciones",
                "operationId": "32430190a83b3e7d02fd0b80ff16641f",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las inscripciones."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva inscripción",
                "operationId": "be6247395db6a4f769454770d4c58d03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "alumno_id",
                                    "clase_id",
                                    "materia_id",
                                    "estado_pago"
                                ],
                                "properties": {
                                    "alumno_id": {
                                        "type": "integer"
                                    },
                                    "clase_id": {
                                        "type": "integer"
                                    },
                                    "materia_id": {
                                        "type": "integer"
                                    },
                                    "estado_pago": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscripción creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/inscripciones/{id}": {
            "get": {
                "summary": "Mostrar detalle de una inscripción",
                "operationId": "b4666d48f6ff278358ab8b6da9e61464",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la inscripción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar detalle de una inscripción."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "summary": "Actualizar una inscripción existente",
                "operationId": "ba064dcee1714f45ba894b148f20757f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la inscripción a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "alumno_id",
                                    "clase_id",
                                    "materia_id",
                                    "estado_pago"
                                ],
                                "properties": {
                                    "alumno_id": {
                                        "type": "integer"
                                    },
                                    "clase_id": {
                                        "type": "integer"
                                    },
                                    "materia_id": {
                                        "type": "integer"
                                    },
                                    "estado_pago": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscripción actualizada exitosamente"
                    },
                    "404": {
                        "description": "Inscripción no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una inscripción existente",
                "operationId": "bfdeb5bc6dc0090e3f0b828f5671c055",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la inscripción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inscripción eliminada exitosamente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/materias": {
            "get": {
                "summary": "Mostrar todas las materias",
                "operationId": "918cb2607842cfe87ff9f6796c4cc540",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las materias."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva materia",
                "operationId": "6d2751d41fdddb317afd753ab2d78f1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "nivel"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "nivel": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Materia creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/materias/{id}": {
            "get": {
                "summary": "Mostrar detalle de una materia",
                "operationId": "808b887ba424ec0048dbd9a11aa22ffe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar detalle de una materia."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "summary": "Actualizar una materia existente",
                "operationId": "b2823e9405096afc917077879051c33e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "nivel"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "nivel": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materia actualizada exitosamente"
                    },
                    "404": {
                        "description": "Materia no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una materia existente",
                "operationId": "aa3b33b4b37a772aa16e47d56f319567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materia eliminada exitosamente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}